feat: Improve user follow functionality
Add the ability to follow users in the Userlist component. This allows users to follow other users on "The Social" social media platform for developers. The followUser function in the useFollow hook is used to update the current user's following list in the database. This feature enhances the social interaction and networking capabilities of the platform.